Fix fallback to default namepsace when namespace passed as an option
Falling back to default namespace works ok with namespace as a prefix because the prefix is stripped as part of the pass through `_find` in the first instance. However if the namespace is defined as part of the options parameter then it is still present on the options for the second (fallback) call to `_find`.

Fix by setting the namespace explicitly to the default namespace when passing through to the fallback call to `_find`.